pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyMaekRetrieval"
version = "1.2.0"
authors = [
  { name="Andreas Foth", email="author@example.com" },
]
description = "Python package for making microwave radiometer retrievals"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "numpy",
  "xarray",
  "netcdf4",
  "scikit-learn",
  "pyyaml",
  "matplotlib",
]

[project.urls]
"Homepage" = "https://github.com/remsens-lim/pyMakeRetrieval"
"Bug Tracker" = "https://github.com/remsens-lim/pyMakeRetrieval/issues"
